+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
  - Issue #6
  - Issue #8
+ - Some functions using "DEFAULTS" as a key instead of "DEFAULT"
 
 ## [1.1] - 2021-01-07
 ### Added